fixed startup bugs when missing adelante or abi file
