+++ src/messages.ts
@@ -1,4 +1,3 @@
-
 export const initGreeting = () => {
   console.log("");
   console.log("");
@@ -7,11 +6,12 @@ export const initGreeting = () => {
   console.log("Welcome to");
   console.log("");
   console.log(
-`
+    `
 ░█▀█░█▀▄░█▀▀░█░░░█▀█░█▀█░▀█▀░█▀▀
 ░█▀█░█░█░█▀▀░█░░░█▀█░█░█░░█░░█▀▀
 ░▀░▀░▀▀░░▀▀▀░▀▀▀░▀░▀░▀░▀░░▀░░▀▀▀
-`);
+`
+  );
   console.log("");
   console.log("React code generation for Solidity!");
   console.log("");
@@ -20,30 +20,31 @@ export const initGreeting = () => {
   console.log("<:><:><:><:><:><:><:><:><:><:><:>");
   console.log("");
   console.log("");
-}
+};
 
 export const generatorGreeting = () => {
   console.log("");
   console.log("<:><:><:><:><:><:><:><:><:><:><:>");
   console.log("");
-console.log(
-`
+  console.log(
+    `
 ░█▀█░█▀▄░█▀▀░█░░░█▀█░█▀█░▀█▀░█▀▀
 ░█▀█░█░█░█▀▀░█░░░█▀█░█░█░░█░░█▀▀
 ░▀░▀░▀▀░░▀▀▀░▀▀▀░▀░▀░▀░▀░░▀░░▀▀▀
-`);
+`
+  );
   console.log("        Created by Seyaji  ");
   console.log("                           v1.1.5");
   console.log("<:><:><:><:><:><:><:><:><:><:><:>");
-}
+};
 
 export const inProgress = () => {
   console.log("");
   console.log("     👻  Generating files  👻");
   console.log("");
-}
+};
 
-export const generatorComplete = () => {  
+export const generatorComplete = () => {
   console.log("        👻  Complete!  👻");
   console.log("");
   console.log("<:><:><:><:><:><:><:><:><:><:><:>");
@@ -60,3 +61,35 @@ export const setupComplete = () => {
   console.log("");
 };
 
+export const missingAbi = () => {
+  console.log("");
+  console.log("<:><:><:><:><:><:><:><:><:><:><:>");
+  console.log("");
+  console.log("");
+  console.log("failed to generate, check the path to your abi.json file");
+  console.log("");
+  console.log("");
+  console.log("<:><:><:><:><:><:><:><:><:><:><:>");
+};
+
+export const unknownFailure = () => {
+  console.log("");
+  console.log("<:><:><:><:><:><:><:><:><:><:><:>");
+  console.log("");
+  console.log("");
+  console.log("adelante failed to generate  files");
+  console.log("");
+  console.log("");
+  console.log("<:><:><:><:><:><:><:><:><:><:><:>");
+};
+
+export const missingAdelante = () => {
+  console.log("<:><:><:><:><:><:><:><:><:><:><:>");
+  console.log("");
+  console.log("Failed to generate...");
+  console.log("Make sure your adelante.json file is configured properly.");
+  console.log("");
+  console.log("In case of type errors please raise an issue on GitHub :)");
+  console.log("");
+  console.log("<:><:><:><:><:><:><:><:><:><:><:>");
+};
