+++ src/utils.ts
@@ -1,11 +1,8 @@
-import appRoot from "app-root-path";
 
 export function capitalize(name: string) {
   return name.charAt(0).toUpperCase() + name.slice(1);
 }
 
-import { createRequire } from "module";
-export const require = createRequire(import.meta.url);
 
 export const useTs = (typescript: boolean, ts: string, js: string) => typescript ? ts : js;
 
@@ -18,16 +15,6 @@ export const propsFilter = (props: any, outputs: any) => (props.filter((prop: an
   }
 }))
 
-export function importer() {
-  try {
-    const adelante = require(`${appRoot}/adelante.json`);
-    console.log(adelante)
-    return adelante;
-  } catch (error) {
-    console.log('trying to import from: ', appRoot + '/adelante.json');
-    console.log(error)
-  }
-}
 
 export const funcFormat = (name: string) => {
   return name.split("").reduce((acc, curr) => {
